requirements.txt
click==7.1.2
colorama==0.4.3
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
python-dotenv==0.14.0
qrcode-artistic==2.1.0
segno==1.3.1
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
